NaviLens_API/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.30
pydantic==2.7.1
python-multipart==0.0.9
Pillow==10.3.0
transformers==4.41.2
torch>=2.2.0
gTTS==2.5.1
ultralytics